Moved repository definitions into a Raunstrup.Data project, and the in-memory solution into a Raunstrup.Data.InMemory project.
This was neccesary as Raunstrup.Core depends on the interfaces **and** implementation, while Raunstrup.Data.InMemory (or any other implementation) also depends on the interfaces. In the old setup, where repository interfaces recided in Raunstrup.Core, this would look like

Raunstrup.Core -> Raunstrup.DataImplementation -> Raunstrup.Core

which is bad, because VS says so. No but really, it's needed if we want to seperate the data access implementation into a seperate project..which I very much want to. Screw you guys. I'm kidding I love you.